Refactor tracing: Move trace support to its own integration
