+++ homeassistant/components/automation/manifest.json
@@ -2,7 +2,7 @@
   "domain": "automation",
   "name": "Automation",
   "documentation": "https://www.home-assistant.io/integrations/automation",
-  "dependencies": ["blueprint"],
+  "dependencies": ["blueprint", "trace"],
   "after_dependencies": [
     "device_automation",
     "webhook"
