+++ homeassistant/components/trace/const.py
@@ -1,0 +1,4 @@
+"""Shared constants for automation and script tracing and debugging."""
+
+DATA_TRACE = "trace"
+STORED_TRACES = 5  # Stored traces per automation
