+++ homeassistant/components/trace/manifest.json
@@ -1,0 +1,9 @@
+{
+  "domain": "trace",
+  "name": "Trace",
+  "documentation": "https://www.home-assistant.io/integrations/automation",
+  "codeowners": [
+    "@home-assistant/core"
+  ],
+  "quality_scale": "internal"
+}
