+++ tests/components/trace/__init__.py
@@ -1,0 +1,1 @@
+"""The tests for Trace."""
